+++ src/Routes/task.js
@@ -1,75 +1,30 @@
 import express from "express";
 import connection from "../database/config.js";
+import { deleteTask, getTask, getTaskById, postTask, updateTask } from "../Controllers/TaskController.js";
 const router = express.Router();
 
-router.get("/get", async (req, res) => {
-    let query = `SELECT * FROM tasks`;
-    await connection.query(query, (err, result) => {
-        if (err) {
-            console.log("Error executing query:", err.stack);
-        }
-        return res.status(200).json(result);
+const cleanUpTask = () => {
+    let query = "DELETE FROM tasks";
+
+    connection.query(query, (err, result) => {
+        if (err) console.error("Error executing query:", err.stack);
+        console.log({result, message: "Database cleaned up"});
     });
-});
+};
 
-router.get("/get/:id", async (req, res) => {
-    const { id } = req.params;
-    let query = `SELECT * FROM tasks WHERE id = ?`;
-    await connection.query(query, [id], (err, result) => {
-        if (err) {
-            console.log("Error executing query:", err.stack);
-        }
-        return res.status(200).json(result);
-    });
-});
+const startBackgroundTask = () => {
+    setInterval(cleanUpTask, 60 * 1000);
+};
+startBackgroundTask();
 
-router.post("/post", async (req, res) => {
-    const requestBody = {
-        title: req.body.title,
-        description: req.body.description,
-        due_date: req.body.due_date,
-    };
+router.get("/get", getTask);
 
-    let query = `INSERT INTO tasks (title, description, due_date) VALUES (?, ?, ?)`;
-    await connection.query(
-        query,
-        [requestBody.title, requestBody.description, requestBody.due_date],
-        (err, result) => {
-            if (err) {
-                console.error("Error executing query:", err.stack);
-            } else {
-                res.status(200).json({ message: "inserted correctly" });
-                return console.log(result);
-            }
-        }
-    );
-});
+router.get("/get/:id", getTaskById);
 
-router.put("/put/:id", async (req, res) => {
-    const { id } = req.params;
-    const { title, description } = req.body;
-    console.log({ title, description });
-    let query = `UPDATE tasks set title=?, description= ? WHERE id = ?`;
-    connection.query(query, [ title, description, id], (err, result) => {
-        if (err) {
-            console.log("Error executing query:", err.stack);
-        }
-        return res.status(200).json({ message: "Element modified" });
-    });
-});
+router.post("/post", postTask);
 
-router.delete("/delete/:id", async (req, res) => {
-    const { id } = req.params;
+router.put("/put/:id", updateTask);
 
-    let sql = `DELETE FROM tasks WHERE id = ?`;
-    await connection.query(sql, [id], function (err, result) {
-        if (err) {
-            console.error("Error executing query:", err.stack);
-        } else {
-            res.status(200).json({ message: "deleted correctly" });
-            return console.log(result);
-        }
-    });
-});
+router.delete("/delete/:id", deleteTask);
 
 export default router;
